Built a new Sign up and login page.
Visitors will now hi9t the splash page which offers them a chance to signup or view the website without loging in.
If they do view as a visitor (guest) then they will not see a user name in the top right but a signup button.

This will bring them to the login page.

Need to add "Not a member? Sign up here functionallity"

Please note that none of this uses rigorous authentication. This trusts that the cookie has full integrity.
This is simply because I wanted to make the work flow of the site and that part of the functionality first so some aspects of it are present.